Creating Accordion component aside from peekFont
